+++ CHANGELOG.md
@@ -4,8 +4,19 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
-### Changed
-- check-puppet-last-run.rb: Added critical alerting on failures during summary file processing
+### Added
+- Testing for Ruby 2.4.1
+
+## [1.2.0] - 2017-05-30
+### Added
+- `check-puppet-last-run.rb`: Added option for reporting failed restarts (@antonidabek)
+
+## [1.1.0] - 2017-01-30
+### Added
+- `check-puppet-last-run.rb`: Added critical alerting on failures during summary file processing (@grem11n)
+
+### Fixed
+- `check-puppet-last-run.rb`: Corrected failure count hash key (@eahola)
 
 ## [1.0.0] - 2016-06-21
 ### Changed
@@ -26,6 +37,8 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ### Added
 - initial release
 
-[Unreleased]: https://github.com/sensu-plugins/sensu-plugins-puppet/compare/1.0.0...HEAD
+[Unreleased]: https://github.com/sensu-plugins/sensu-plugins-puppet/compare/1.2.0...HEAD
+[1.2.0]: https://github.com/sensu-plugins/sensu-plugins-puppet/compare/1.1.0...1.2.0
+[1.1.0]: https://github.com/sensu-plugins/sensu-plugins-puppet/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/sensu-plugins/sensu-plugins-puppet/compare/0.0.2...1.0.0
 [0.0.2]: https://github.com/sensu-plugins/sensu-plugins-puppet/compare/0.0.1...0.0.2